+++ tiger-test-lib/src/main/java/de/gematik/test/tiger/lib/TigerDirector.java
@@ -91,7 +91,7 @@ public class TigerDirector {
         log.info("Registering shutdown hook...");
         Runtime.getRuntime().addShutdownHook(new Thread(() -> {
             try {
-                if (getLibConfig().isActivateWorkflowUi()) {
+                if (getLibConfig().isActivateWorkflowUi() && !tigerTestEnvMgr.isUserAcknowledgedShutdown()) {
                     System.out.println(Ansi.colorize("TGR Workflow UI is active, please press quit in browser window...", RbelAnsiColors.GREEN_BOLD));
                     if (tigerTestEnvMgr != null) {
                         tigerTestEnvMgr.receiveTestEnvUpdate(TigerStatusUpdate.builder()
@@ -256,12 +256,23 @@ public class TigerDirector {
     }
 
     public static void waitForQuit() {
-        // TODO TGR-516 wait for quit from workflow ui, then shutdown env and all processes and system exit to force mvn failsafe to abort ....
-        envMgrApplicationContext.close();
-        await()
-            .pollInterval(Duration.ofMillis(100))
-            .atMost(Duration.ofDays(1))
-            .until(() -> !envMgrApplicationContext.isRunning());
+        if (getLibConfig().isActivateWorkflowUi()) {
+            tigerTestEnvMgr.receiveTestEnvUpdate(TigerStatusUpdate.builder()
+                .bannerMessage("Press QUIT to abort test run")
+                .bannerColor("green")
+                .bannerType(BannerType.TESTRUN_ENDED)
+                .build());
+            try {
+                await().pollInterval(1, TimeUnit.SECONDS)
+                    .atMost(5, TimeUnit.HOURS)
+                    .until(() -> tigerTestEnvMgr.isUserAcknowledgedShutdown());
+            } finally {
+                System.exit(0);
+            }
+        } else {
+            tigerTestEnvMgr.waitForConsoleInput("quit");
+            System.exit(0);
+        }
     }
 
     private final static Pattern showSteps = Pattern.compile(".*TGR (zeige|show) ([\\w|ü|ß]*) (Banner|banner|text|Text) \"(.*)\"");//NOSONAR
@@ -317,5 +328,21 @@ public class TigerDirector {
             SerenityRest.replaceFiltersWith(new ArrayList<>());
         }
         curlLoggingFilter = null;
+    }
+
+    public static void pauseExecution() {
+        if (getLibConfig().isActivateWorkflowUi()) {
+            tigerTestEnvMgr.receiveTestEnvUpdate(TigerStatusUpdate.builder()
+                .bannerMessage("Test execution paused, click to continue")
+                .bannerColor("green")
+                .bannerType(BannerType.STEP_WAIT)
+                .build());
+            await().pollInterval(1, TimeUnit.SECONDS)
+                .atMost(5, TimeUnit.HOURS)
+                .until(() -> tigerTestEnvMgr.isUserAcknowledgedContinueTestRun());
+            tigerTestEnvMgr.resetUserAcknowledgedContinueTestRun();
+        } else {
+            tigerTestEnvMgr.waitForConsoleInput("next");
+        }
     }
 }