+++ tiger-test-lib/src/test/java/de/gematik/test/tiger/TestTigerCucumberListener.java
@@ -5,7 +5,9 @@
 package de.gematik.test.tiger;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import de.gematik.test.tiger.common.config.TigerGlobalConfiguration;
 import de.gematik.test.tiger.lib.TigerDirector;
+import de.gematik.test.tiger.testenvmgr.TigerTestEnvMgr;
 import de.gematik.test.tiger.testenvmgr.controller.EnvStatusController;
 import de.gematik.test.tiger.testenvmgr.data.TigerEnvStatusDto;
 import de.gematik.test.tiger.testenvmgr.env.ScenarioUpdate;
@@ -49,6 +51,7 @@ public class TestTigerCucumberListener {
     @BeforeEach
     public void initListener() {
         listener = new TigerCucumberListener();
+        TigerGlobalConfiguration.putValue(TigerTestEnvMgr.CFG_PROP_NAME_LOCAL_PROXY_ADMIN_PORT, 9999);
         envStatusController.getStatus().getFeatureMap().clear();
     }